fix: import umd instead of sources
Jest acts up when I import store sources from ui-context.
Switching to the umd fixes it.

I think the difference with the api-client which imports correctly
when I add it to ignoreTransform regex in jest config
is that the context library is built as a vue-cli library and does not
export typings.